chore(server): move rich text normalize upgrade command to 2-37 (current version)
