+++ packages/twenty-server/src/database/commands/upgrade-version-command/2-37/utils/normalize-record-crud-rich-text-fields.util.ts
