+++ packages/twenty-server/src/database/commands/upgrade-version-command/workspace-command-provider.module.ts
@@ -29,6 +29,7 @@ import { V2_33_UpgradeVersionCommandModule } from 'src/database/commands/upgrade
 import { V2_34_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-version-command/2-34/2-34-upgrade-version-command.module';
 import { V2_35_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-version-command/2-35/2-35-upgrade-version-command.module';
 import { V2_36_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-version-command/2-36/2-36-upgrade-version-command.module';
+import { V2_37_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-version-command/2-37/2-37-upgrade-version-command.module';
 import { V2_4_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-version-command/2-4/2-4-upgrade-version-command.module';
 import { V2_5_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-version-command/2-5/2-5-upgrade-version-command.module';
 import { V2_7_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-version-command/2-7/2-7-upgrade-version-command.module';
@@ -71,6 +72,7 @@ import { V2_9_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-
     V2_34_UpgradeVersionCommandModule,
     V2_35_UpgradeVersionCommandModule,
     V2_36_UpgradeVersionCommandModule,
+    V2_37_UpgradeVersionCommandModule,
   ],
 })
 export class WorkspaceCommandProviderModule {}
